deleting dishes and employees

+++ src/main/java/com/dev/restaurant/dao/DishDao.java
@@ -6,7 +6,10 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.Optional;
 
 
-public interface DishDao extends CrudRepository<Dish, Integer> {
+public interface DishDao extends CrudRepository<Dish, Long> {
 
     Optional<Dish> findByName(String name);
+
+    //Optional<Dish> findById(Long id);
+
 }
